+++ http4s-quickstart/src/main/scala/com/github/kkrull/http4s/greet/HelloWorld.scala
@@ -1,4 +1,4 @@
-package com.github.kkrull.http4s
+package com.github.kkrull.http4s.greet
 
 import cats.Applicative
 import cats.implicits._
