+++ http4s-quickstart/src/main/scala/com/github/kkrull/http4s/joke/JokeRoutes.scala
@@ -1,0 +1,20 @@
+package com.github.kkrull.http4s.joke
+
+import cats.effect.Sync
+import cats.implicits._
+import org.http4s.HttpRoutes
+import org.http4s.dsl.Http4sDsl
+
+object JokeRoutes {
+  def jokeRoutes[F[_]: Sync](J: Jokes[F]): HttpRoutes[F] = {
+    val dsl = new Http4sDsl[F] {}
+
+    import dsl._
+    HttpRoutes.of[F] { case GET -> Root / "joke" =>
+      for {
+        joke <- J.get
+        resp <- Ok(joke)
+      } yield resp
+    }
+  }
+}
