+++ http4s-quickstart/src/main/scala/com/github/kkrull/http4s/joke/Jokes.scala
@@ -1,4 +1,4 @@
-package com.github.kkrull.http4s
+package com.github.kkrull.http4s.joke
 
 import cats.effect.Concurrent
 import cats.implicits._
